requirements.txt
Django==1.10
django-filter==0.14.0
djangorestframework==3.4.6
Markdown==2.6.6
pkg-resources==0.0.0
